+++ packages/google_maps_flutter/android/src/main/java/io/flutter/plugins/googlemaps/GoogleMapsPlugin.java
@@ -6,13 +6,15 @@ package io.flutter.plugins.googlemaps;
 
 import android.app.Activity;
 import android.app.Application;
-import android.app.FragmentManager;
 import android.os.Bundle;
-import com.google.android.gms.maps.GoogleMap;
-import com.google.android.gms.maps.MapFragment;
-import com.google.android.gms.maps.OnMapReadyCallback;
-import io.flutter.plugin.common.MethodCall;
-import io.flutter.plugin.common.MethodChannel;
+import androidx.annotation.NonNull;
+import androidx.lifecycle.DefaultLifecycleObserver;
+import androidx.lifecycle.Lifecycle;
+import androidx.lifecycle.LifecycleOwner;
+import io.flutter.embedding.engine.plugins.FlutterPlugin;
+import io.flutter.embedding.engine.plugins.activity.ActivityAware;
+import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding;
+import io.flutter.embedding.engine.plugins.lifecycle.FlutterLifecycleAdapter;
 import io.flutter.plugin.common.PluginRegistry.Registrar;
 import java.util.concurrent.atomic.AtomicInteger;
 
@@ -23,7 +25,10 @@ import java.util.concurrent.atomic.AtomicInteger;
  * overlay.
  */
 public class GoogleMapsPlugin
-    implements Application.ActivityLifecycleCallbacks, MethodChannel.MethodCallHandler {
+    implements Application.ActivityLifecycleCallbacks,
+        FlutterPlugin,
+        ActivityAware,
+        DefaultLifecycleObserver {
   static final int CREATED = 1;
   static final int STARTED = 2;
   static final int RESUMED = 3;
@@ -31,8 +36,11 @@ public class GoogleMapsPlugin
   static final int STOPPED = 5;
   static final int DESTROYED = 6;
   private final AtomicInteger state = new AtomicInteger(0);
-  private final int registrarActivityHashCode;
-  private final FragmentManager fragmentManager;
+  private int registrarActivityHashCode;
+  private FlutterPluginBinding pluginBinding;
+  private Lifecycle lifecycle;
+
+  private static final String VIEW_TYPE = "plugins.flutter.io/google_maps";
 
   public static void registerWith(Registrar registrar) {
     if (registrar.activity() == null) {
@@ -40,13 +48,97 @@ public class GoogleMapsPlugin
       // We stop the registration process as this plugin is foreground only.
       return;
     }
-    final GoogleMapsPlugin plugin = new GoogleMapsPlugin(registrar);
+    final GoogleMapsPlugin plugin = new GoogleMapsPlugin(registrar.activity());
     registrar.activity().getApplication().registerActivityLifecycleCallbacks(plugin);
     registrar
         .platformViewRegistry()
         .registerViewFactory(
-            "plugins.flutter.io/google_maps", new GoogleMapFactory(plugin.state, registrar));
+            VIEW_TYPE,
+            new GoogleMapFactory(plugin.state, registrar.messenger(), null, null, registrar, -1));
   }
+
+  public GoogleMapsPlugin() {}
+
+  // FlutterPlugin
+
+  @Override
+  public void onAttachedToEngine(FlutterPluginBinding binding) {
+    pluginBinding = binding;
+  }
+
+  @Override
+  public void onDetachedFromEngine(FlutterPluginBinding binding) {
+    pluginBinding = null;
+  }
+
+  // ActivityAware
+
+  @Override
+  public void onAttachedToActivity(ActivityPluginBinding binding) {
+    lifecycle = FlutterLifecycleAdapter.getActivityLifecycle(binding);
+    lifecycle.addObserver(this);
+    pluginBinding
+        .getPlatformViewRegistry()
+        .registerViewFactory(
+            VIEW_TYPE,
+            new GoogleMapFactory(
+                state,
+                pluginBinding.getBinaryMessenger(),
+                binding.getActivity().getApplication(),
+                lifecycle,
+                null,
+                binding.getActivity().hashCode()));
+  }
+
+  @Override
+  public void onDetachedFromActivity() {
+    lifecycle.removeObserver(this);
+  }
+
+  @Override
+  public void onDetachedFromActivityForConfigChanges() {
+    this.onDetachedFromActivity();
+  }
+
+  @Override
+  public void onReattachedToActivityForConfigChanges(ActivityPluginBinding binding) {
+    lifecycle = FlutterLifecycleAdapter.getActivityLifecycle(binding);
+    lifecycle.addObserver(this);
+  }
+
+  // DefaultLifecycleObserver methods
+
+  @Override
+  public void onCreate(@NonNull LifecycleOwner owner) {
+    state.set(CREATED);
+  }
+
+  @Override
+  public void onStart(@NonNull LifecycleOwner owner) {
+    state.set(STARTED);
+  }
+
+  @Override
+  public void onResume(@NonNull LifecycleOwner owner) {
+    state.set(RESUMED);
+  }
+
+  @Override
+  public void onPause(@NonNull LifecycleOwner owner) {
+    state.set(PAUSED);
+  }
+
+  @Override
+  public void onStop(@NonNull LifecycleOwner owner) {
+    state.set(STOPPED);
+  }
+
+  @Override
+  public void onDestroy(@NonNull LifecycleOwner owner) {
+    state.set(DESTROYED);
+  }
+
+  // Application.ActivityLifecycleCallbacks methods
 
   @Override
   public void onActivityCreated(Activity activity, Bundle savedInstanceState) {
@@ -100,6 +192,8 @@ public class GoogleMapsPlugin
     state.set(DESTROYED);
   }
 
+  private GoogleMapsPlugin(Activity activity) {
+    this.registrarActivityHashCode = activity.hashCode();
   private GoogleMapsPlugin(Registrar registrar) {
     Activity activity = registrar.activity();
     registrarActivityHashCode = activity.hashCode();