+++ packages/google_maps_flutter/CHANGELOG.md
@@ -1,6 +1,12 @@
-## 0.5.22
+## 0.5.23
 
 * Android: Add dummy map initialization
+
+## 0.5.22
+
+* Support Android v2 embedding.
+* Bump the min flutter version to `1.12.13+hotfix.5`.
+* Fixes some e2e tests on Android.
 
 ## 0.5.21+17